move flowCell to its own directory. import templateOf
