+++ src/flowCell/templateOf.js
